ref: factory functions are defined outside factory.ts
